Restructure new classes
